backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
requests>=2.31
urllib3<2
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
django-redis>=5.4
